+++ LabModule/admin.py
@@ -2,22 +2,22 @@
 
 from django.contrib import admin
 
-from models import Bandeja
-from models import Experimento
-from models import LaboratorioProfile
-from models import LugarAlmacenamiento
-from models import LugarAlmacenamientoEnLab
-from models import MaquinaEnLab
-from models import MaquinaProfile
-from models import MaquinaSolicitud
-from models import Muestra
-from models import MuestraSolicitud
-from models import Paso
-from models import Protocolo
-from models import Proyecto
-from models import Solicitud
-from models import TipoDocumento
-from models import Usuario
+from LabModule.app_models.Almacenamiento import Almacenamiento
+from LabModule.app_models.AlmacenamientoEnLab import AlmacenamientoEnLab
+from LabModule.app_models.Bandeja import Bandeja
+from LabModule.app_models.Experimento import Experimento
+from LabModule.app_models.Laboratorio import Laboratorio
+from LabModule.app_models.Maquina import Maquina
+from LabModule.app_models.MaquinaEnLab import MaquinaEnLab
+from LabModule.app_models.Muestra import Muestra
+from LabModule.app_models.Paso import Paso
+from LabModule.app_models.Protocolo import Protocolo
+from LabModule.app_models.Proyecto import Proyecto
+from LabModule.app_models.Solicitud import Solicitud
+from LabModule.app_models.SolicitudMaquina import SolicitudMaquina
+from LabModule.app_models.SolicitudMuestra import SolicitudMuestra
+from LabModule.app_models.TipoDocumento import TipoDocumento
+from LabModule.app_models.Usuario import Usuario
 
 
 class TipoDocumentoAdmin(admin.ModelAdmin):
@@ -35,19 +35,19 @@ class UsuarioAdmin(admin.ModelAdmin):
 
 
 # Register your models here.
-admin.site.register(Experimento)
-admin.site.register(LaboratorioProfile)
-admin.site.register(MaquinaEnLab)
-admin.site.register(MaquinaProfile)
-admin.site.register(Muestra)
-admin.site.register(MuestraSolicitud)
-admin.site.register(Paso)
-admin.site.register(Proyecto)
-admin.site.register(Protocolo)
-admin.site.register(Solicitud)
+admin.site.register(Almacenamiento)
+admin.site.register(AlmacenamientoEnLab)
 admin.site.register(Bandeja)
-admin.site.register(LugarAlmacenamientoEnLab)
-admin.site.register(MaquinaSolicitud)
-admin.site.register(LugarAlmacenamiento)
+admin.site.register(Experimento)
+admin.site.register(Laboratorio)
+admin.site.register(Maquina)
+admin.site.register(MaquinaEnLab)
+admin.site.register(Muestra)
+admin.site.register(Paso)
+admin.site.register(Protocolo)
+admin.site.register(Proyecto)
+admin.site.register(Solicitud)
+admin.site.register(SolicitudMaquina)
+admin.site.register(SolicitudMuestra)
 admin.site.register(TipoDocumento, TipoDocumentoAdmin)
 admin.site.register(Usuario, UsuarioAdmin)
