+++ LabModule/app_models/__init__.py
@@ -1,0 +1,16 @@
+from .Almacenamiento import Almacenamiento
+from .AlmacenamientoEnLab import AlmacenamientoEnLab
+from .Bandeja import Bandeja
+from .Experimento import Experimento
+from .Laboratorio import Laboratorio
+from .Maquina import Maquina
+from .MaquinaEnLab import MaquinaEnLab
+from .Muestra import Muestra
+from .Paso import Paso
+from .Protocolo import Protocolo
+from .Proyecto import Proyecto
+from .Solicitud import Solicitud
+from .SolicitudMaquina import SolicitudMaquina
+from .SolicitudMuestra import SolicitudMuestra
+from .TipoDocumento import TipoDocumento
+from .Usuario import Usuario
