+++ LabModule/management/commands/creardatosaplicacion.py
@@ -12,75 +12,84 @@ from django.core.files.temp import NamedTemporaryFile
 from django.core.management.base import BaseCommand
 from django.templatetags.static import static
 
-from LabModule.models import Bandeja
-from LabModule.models import Experimento
-from LabModule.models import LaboratorioProfile
-from LabModule.models import LugarAlmacenamiento
-from LabModule.models import LugarAlmacenamientoEnLab
-from LabModule.models import MaquinaEnLab
-from LabModule.models import MaquinaProfile
-from LabModule.models import Muestra
-from LabModule.models import Paso
-from LabModule.models import Protocolo
-from LabModule.models import Proyecto
-from LabModule.models import TipoDocumento
-from LabModule.models import Usuario
+from LabModule.app_models.Almacenamiento import Almacenamiento
+from LabModule.app_models.AlmacenamientoEnLab import AlmacenamientoEnLab
+from LabModule.app_models.Bandeja import Bandeja
+from LabModule.app_models.Experimento import Experimento
+from LabModule.app_models.Laboratorio import Laboratorio
+from LabModule.app_models.MaquinaEnLab import MaquinaEnLab
+from LabModule.app_models.Muestra import Muestra
+from LabModule.app_models.Paso import Paso
+from LabModule.app_models.Protocolo import Protocolo
+from LabModule.app_models.Proyecto import Proyecto
+from LabModule.app_models.TipoDocumento import TipoDocumento
+from LabModule.app_models.Usuario import Usuario
 
 SUPERUSUARIO = getattr(settings, "SUPERUSUARIO", 'admin')
 CONTRASENA = getattr(settings, "CONTRASENA", '1a2d3m4i5n6')
 EMAIL_HOST_USER = getattr(settings, "EMAIL_HOST_USER", 'admin@admin.com')
 
-can_addProject = Permission.objects.get(name='proyecto||agregar')
-can_editProject = Permission.objects.get(name='proyecto||editar')
-can_listProject = Permission.objects.get(name='proyecto||listar')
-can_viewProject = Permission.objects.get(name='proyecto||ver')
-can_addExperiment = Permission.objects.get(name='experimento||agregar')
-can_editExperiment = Permission.objects.get(name='experimento||editar')
-can_listExperiment = Permission.objects.get(name='experimento||listar')
-can_viewExperiment = Permission.objects.get(name='experimento||ver')
-can_addUser = Permission.objects.get(name='usuario||agregar')
-can_editUser = Permission.objects.get(name='usuario||editar')
-can_listUser = Permission.objects.get(name='usuario||listar')
-can_viewUser = Permission.objects.get(name='usuario||ver')
-can_addLab = Permission.objects.get(name='laboratorio||agregar')
-can_editLab = Permission.objects.get(name='laboratorio||editar')
-can_listLab = Permission.objects.get(name='laboratorio||listar')
-can_viewLab = Permission.objects.get(name='laboratorio||ver')
-can_addMachine = Permission.objects.get(name='maquina||agregar')
-can_editMachine = Permission.objects.get(name='maquina||editar')
-can_listMachine = Permission.objects.get(name='maquina||listar')
-can_viewMachine = Permission.objects.get(name='maquina||ver')
-can_requestMachine = Permission.objects.get(name='maquina||solicitar')
-can_addStorage = Permission.objects.get(name='almacenamiento||agregar')
-can_editStorage = Permission.objects.get(name='almacenamiento||editar')
-can_listStorage = Permission.objects.get(name='almacenamiento||listar')
-can_viewStorage = Permission.objects.get(name='almacenamiento||ver')
-can_requestStorage = Permission.objects.get(name='almacenamiento||solicitar')
-can_addProtocol = Permission.objects.get(name='protocolo||agregar')
-can_editProtocol = Permission.objects.get(name='protocolo||editar')
-can_listProtocol = Permission.objects.get(name='protocolo||listar')
-can_viewProtocol = Permission.objects.get(name='protocolo||ver')
-can_addStep = Permission.objects.get(name='paso||agregar')
-can_editStep = Permission.objects.get(name='paso||editar')
-can_listStep = Permission.objects.get(name='paso||listar')
-can_viewStep = Permission.objects.get(name='paso||ver')
-can_addSample = Permission.objects.get(name='muestra||agregar')
-can_editSample = Permission.objects.get(name='muestra||editar')
-can_listSample = Permission.objects.get(name='muestra||listar')
-can_viewSample = Permission.objects.get(name='muestra||ver')
-can_requestSample = Permission.objects.get(name='muestra||solicitar')
-can_listRequest = Permission.objects.get(name='solicitud||listar')
-can_viewRequest = Permission.objects.get(name='solicitud||ver')
-can_manageRequest = Permission.objects.get(name='solicitud||admin')
-can_addTray = Permission.objects.get(name='bandeja||agregar')
-can_editTray = Permission.objects.get(name='bandeja||editar')
-can_listTray = Permission.objects.get(name='bandeja||listar')
-can_viewTray = Permission.objects.get(name='bandeja||ver')
+can_addProject = Permission.objects.get(name = 'proyecto||agregar')
+can_editProject = Permission.objects.get(name = 'proyecto||editar')
+can_listProject = Permission.objects.get(name = 'proyecto||listar')
+can_viewProject = Permission.objects.get(name = 'proyecto||ver')
+
+can_addExperiment = Permission.objects.get(name = 'experimento||agregar')
+can_editExperiment = Permission.objects.get(name = 'experimento||editar')
+can_listExperiment = Permission.objects.get(name = 'experimento||listar')
+can_viewExperiment = Permission.objects.get(name = 'experimento||ver')
+
+can_addUser = Permission.objects.get(name = 'usuario||agregar')
+can_editUser = Permission.objects.get(name = 'usuario||editar')
+can_listUser = Permission.objects.get(name = 'usuario||listar')
+can_viewUser = Permission.objects.get(name = 'usuario||ver')
+
+can_addLab = Permission.objects.get(name = 'laboratorio||agregar')
+can_editLab = Permission.objects.get(name = 'laboratorio||editar')
+can_listLab = Permission.objects.get(name = 'laboratorio||listar')
+can_viewLab = Permission.objects.get(name = 'laboratorio||ver')
+
+can_addMachine = Permission.objects.get(name = 'maquina||agregar')
+can_editMachine = Permission.objects.get(name = 'maquina||editar')
+can_listMachine = Permission.objects.get(name = 'Maquina||listar')
+can_viewMachine = Permission.objects.get(name = 'maquina||ver')
+can_requestMachine = Permission.objects.get(name = 'maquina||solicitar')
+
+can_addStorage = Permission.objects.get(name = 'almacenamiento||agregar')
+can_editStorage = Permission.objects.get(name = 'almacenamiento||editar')
+can_listStorage = Permission.objects.get(name = 'almacenamiento||listar')
+can_viewStorage = Permission.objects.get(name = 'almacenamiento||ver')
+can_requestStorage = Permission.objects.get(name = 'almacenamiento||solicitar')
+
+can_addProtocol = Permission.objects.get(name = 'protocolo||agregar')
+can_editProtocol = Permission.objects.get(name = 'protocolo||editar')
+can_listProtocol = Permission.objects.get(name = 'protocolo||listar')
+can_viewProtocol = Permission.objects.get(name = 'protocolo||ver')
+
+can_addStep = Permission.objects.get(name = 'paso||agregar')
+can_editStep = Permission.objects.get(name = 'paso||editar')
+can_listStep = Permission.objects.get(name = 'paso||listar')
+can_viewStep = Permission.objects.get(name = 'paso||ver')
+
+can_addSample = Permission.objects.get(name = 'muestra||agregar')
+can_editSample = Permission.objects.get(name = 'muestra||editar')
+can_listSample = Permission.objects.get(name = 'muestra||listar')
+can_viewSample = Permission.objects.get(name = 'muestra||ver')
+can_requestSample = Permission.objects.get(name = 'muestra||solicitar')
+
+can_listRequest = Permission.objects.get(name = 'solicitud||listar')
+can_viewRequest = Permission.objects.get(name = 'solicitud||ver')
+can_manageRequest = Permission.objects.get(name = 'solicitud||admin')
+
+can_addTray = Permission.objects.get(name = 'bandeja||agregar')
+can_editTray = Permission.objects.get(name = 'bandeja||editar')
+can_listTray = Permission.objects.get(name = 'bandeja||listar')
+can_viewTray = Permission.objects.get(name = 'bandeja||ver')
 
 
 def crearTiposDocumento():
     print ('Creando Tipos de Documento'),
-    nuevoTipoDoc, tipoDocExistente = TipoDocumento.objects.get_or_create(nombre_corto='CC')
+    nuevoTipoDoc, tipoDocExistente = TipoDocumento.objects.get_or_create(nombre_corto = 'CC')
     if tipoDocExistente:
         nuevoTipoDoc.descripcion = 'Cedula de Ciudadania'
         nuevoTipoDoc.save()
@@ -92,12 +101,12 @@ def crearTiposDocumento():
 
 def crearLaboratorio():
     print ('Creando Laboratorios'),
-    LaboratorioProfile.objects.get_or_create(nombre="Laboratorio principal", id="LAB001")
+    Laboratorio.objects.get_or_create(nombre = "Laboratorio principal", id = "LAB001")
     print ('.'),
-    LaboratorioProfile.objects.get_or_create(nombre="Laboratorio Secundario", id="LAB002")
+    Laboratorio.objects.get_or_create(nombre = "Laboratorio Secundario", id = "LAB002")
     print ('.'),
-    laboratioExistente = LaboratorioProfile.objects.get_or_create(nombre="Laboratorio Terciario",
-                                                                            id="LAB003")
+    laboratioExistente = Laboratorio.objects.get_or_create(nombre = "Laboratorio Terciario",
+                                                           id = "LAB003")
     print ('.'),
     if laboratioExistente:
         return 0
@@ -109,15 +118,15 @@ def crearMaquina():
     print ('Creando Maquinas'),
     rta = 1
     with open(".///" + static('lab_static/json/maquinas.json')) as data_file:
-        nuevoLab, laboratioExistente = LaboratorioProfile.objects.get_or_create(nombre="Laboratorio principal",
-                                                                                id="LAB001")
+        nuevoLab, laboratioExistente = Laboratorio.objects.get_or_create(nombre = "Laboratorio principal",
+                                                                         id = "LAB001")
         data = json.load(data_file)
         for maquina in data:
-            nuevaMaquina, maquinaExistente = MaquinaProfile.objects.get_or_create(nombre=maquina['nombre'],
-                                                                                  descripcion=maquina['descripcion'],
-                                                                                  idSistema=maquina['idSistema'],
-                                                                                  con_reserva=maquina['con_reserva']
-                                                                                  )
+            nuevaMaquina, maquinaExistente = maquina.objects.get_or_create(nombre = maquina['nombre'],
+                                                                           descripcion = maquina['descripcion'],
+                                                                           idSistema = maquina['idSistema'],
+                                                                           con_reserva = maquina['con_reserva']
+                                                                           )
             if maquinaExistente:
                 if not maquina['imagen'] == '':
                     img_url = maquina['imagen']
@@ -127,10 +136,10 @@ def crearMaquina():
                     img_temp.flush()
                     nuevaMaquina.imagen.save(img_filename, File(img_temp))
 
-            MaquinaEnLab.objects.get_or_create(idLaboratorio=nuevoLab,
-                                               idMaquina=nuevaMaquina,
-                                               posX=maquina['x'],
-                                               posY=maquina['y'])
+            MaquinaEnLab.objects.get_or_create(idLaboratorio = nuevoLab,
+                                               idMaquina = nuevaMaquina,
+                                               posX = maquina['x'],
+                                               posY = maquina['y'])
             print ('.'),
             if maquinaExistente:
                 rta = 0
@@ -161,7 +170,7 @@ def crearAlmacenamiento():
             posX = row['posX']
             posY = row['posY']
             idLaboratorio = row['idLaboratorio']
-            nuevoLab = LaboratorioProfile.objects.get_or_create(id=idLaboratorio)
+            nuevoLab = Laboratorio.objects.get_or_create(id = idLaboratorio)
             img_url = imagen
             img_filename = urlparse(img_url).path.split('/')[-1]
             img_temp = NamedTemporaryFile()
@@ -169,12 +178,12 @@ def crearAlmacenamiento():
             img_temp.flush()
             for i in range(1, int(cantidad) + 1):
                 idAct += 1
-                nuevoLugar, lugarCreado = LugarAlmacenamiento.objects.get_or_create(nombre=nombre + " " + str(i),
-                                                                                    descripcion=descripcion,
-                                                                                    capacidad=capacidad,
-                                                                                    temperatura=temperatura,
-                                                                                    estado=estado,
-                                                                                    id=idAct)
+                nuevoLugar, lugarCreado = Almacenamiento.objects.get_or_create(nombre = nombre + " " + str(i),
+                                                                               descripcion = descripcion,
+                                                                               capacidad = capacidad,
+                                                                               temperatura = temperatura,
+                                                                               estado = estado,
+                                                                               id = idAct)
                 nuevoLugar.imagen.save(img_filename, File(img_temp))
                 print ('.'),
                 if lugarCreado:
@@ -184,8 +193,8 @@ def crearAlmacenamiento():
                     if xPos > 10:
                         xPos = xPos - 10
                         yPos = yPos + 1
-                    LugarAlmacenamientoEnLab.objects.get_or_create(
-                        idLaboratorio=nuevoLab, idLugar=nuevoLugar, posX=xPos, posY=yPos)
+                    AlmacenamientoEnLab.objects.get_or_create(
+                            idLaboratorio = nuevoLab, idLugar = nuevoLugar, posX = xPos, posY = yPos)
     return rta
 
 
@@ -205,11 +214,11 @@ def crearMuestra():
             imagen = row['imagen']
             unidadBase = row['unidadBase']
             idAlmacenamiento = row['idAlmacenamiento']
-            nuevaMuestra, mustraCreada = Muestra.objects.get_or_create(nombre=nombre, descripcion=descripcion,
-                                                                       valor=valor, activa=activa,
-                                                                       controlado=controlado,
-                                                                       unidadBase=unidadBase)
-            nuevoLugar = LugarAlmacenamiento.objects.get_or_create(id=idAlmacenamiento)
+            nuevaMuestra, mustraCreada = Muestra.objects.get_or_create(nombre = nombre, descripcion = descripcion,
+                                                                       valor = valor, activa = activa,
+                                                                       controlado = controlado,
+                                                                       unidadBase = unidadBase)
+            nuevoLugar = Almacenamiento.objects.get_or_create(id = idAlmacenamiento)
             if mustraCreada:
                 print ('.'),
                 rta = 0
@@ -219,12 +228,12 @@ def crearMuestra():
                 img_temp.write(urllib2.urlopen(img_url).read())
                 img_temp.flush()
                 nuevaMuestra.imagen.save(img_filename, File(img_temp))
-                cuenta = Bandeja.objects.filter(lugarAlmacenamiento=nuevoLugar).count()
+                cuenta = Bandeja.objects.filter(lugarAlmacenamiento = nuevoLugar).count()
                 posicion = 1 if cuenta == 0 else cuenta + 1
-                Bandeja.objects.get_or_create(muestra=nuevaMuestra,
-                                                                               lugarAlmacenamiento=nuevoLugar,
-                                                                               posicion=posicion,
-                                                                               libre=False)
+                Bandeja.objects.get_or_create(muestra = nuevaMuestra,
+                                              lugarAlmacenamiento = nuevoLugar,
+                                              posicion = posicion,
+                                              libre = False)
 
     return rta
 
@@ -234,8 +243,8 @@ def crearProyecto():
     nuevoProyecto, noexistia = Proyecto.objects.get_or_create(nombre = "Colombia Viva")
     nuevoProyecto.descripcion = "Proyecto para sintetizar una droga que reduzca el cansancio"
     nuevoProyecto.objetivo = "Crear NZT"
-    nuevoProyecto.lider = Usuario.objects.get(nombre_usuario='acastro')
-    asistentes = Usuario.objects.all().filter(nombre_usuario__startswith='mgalindo')
+    nuevoProyecto.lider = Usuario.objects.get(nombre_usuario = 'acastro')
+    asistentes = Usuario.objects.all().filter(nombre_usuario__startswith = 'mgalindo')
     asistentes = list(asistentes)
     nuevoProyecto.asistentes.add(*asistentes)
     nuevoProyecto.activo = True
@@ -246,13 +255,13 @@ def crearProyecto():
 def crearExperimento():
     print('Crear Experimento'),
     print ('.'),
-    nuevoExperimento, noexistiaexp = Experimento.objects.get_or_create(nombre="Experimento Colombia Viva")
+    nuevoExperimento, noexistiaexp = Experimento.objects.get_or_create(nombre = "Experimento Colombia Viva")
     nuevoExperimento.descripcion = "Experimento que hace parte de Colombia Viva"
     nuevoExperimento.objetivo = "Crear"
     protocolos = Protocolo.objects.all()
     protocolos = list(protocolos)
     nuevoExperimento.protocolos.add(*protocolos)
-    nuevoExperimento.projecto = Proyecto.objects.get(nombre="Colombia Viva")
+    nuevoExperimento.projecto = Proyecto.objects.get(nombre = "Colombia Viva")
     nuevoExperimento.save()
     return 0
 
@@ -260,7 +269,7 @@ def crearExperimento():
 def crearProtocolo():
     print('Crear Protocolos'),
     print ('.'),
-    nuevoProtocolo, noexistiaproto = Protocolo.objects.get_or_create(nombre="Protocolo Colombia Viva")
+    nuevoProtocolo, noexistiaproto = Protocolo.objects.get_or_create(nombre = "Protocolo Colombia Viva")
     nuevoProtocolo.descripcion = "Protocolo que hace parte de Colombia Viva"
     nuevoProtocolo.objetivo = "Crear"
     nuevoProtocolo.save()
@@ -273,7 +282,7 @@ def crearPaso():
     nuevoPaso, noexistiapaso = Paso.objects.get_or_create(nombre = "Paso Colombia Viva")
     nuevoPaso.descripcion = "Paso que hace parte de Colombia Viva"
     nuevoPaso.objetivo = "Crear"
-    protocolo = Protocolo.objects.get(nombre='Protocolo Colombia Viva')
+    protocolo = Protocolo.objects.get(nombre = 'Protocolo Colombia Viva')
     nuevoPaso.protocolo = protocolo
     nuevoPaso.save()
     return 0
@@ -281,87 +290,87 @@ def crearPaso():
 
 def createGroups():
     print('Crear Grupos'),
-    cientificos, created1 = Group.objects.get_or_create(name='Cientifico Experimentado')
+    cientificos, created1 = Group.objects.get_or_create(name = 'Cientifico Experimentado')
     print ('.'),
-    asistentes, created2 = Group.objects.get_or_create(name='Asistente de Laboratorio')
+    asistentes, created2 = Group.objects.get_or_create(name = 'Asistente de Laboratorio')
     print ('.'),
-    jefes, created3 = Group.objects.get_or_create(name='Jefe de Laboratorio')
+    jefes, created3 = Group.objects.get_or_create(name = 'Jefe de Laboratorio')
     print ('.'),
 
-    Permission.objects.get(name='maquina||ver')
-    Permission.objects.get(name='maquina||solicitar')
-    Permission.objects.get(name='usuario||agregar')
-    Permission.objects.get(name='muestra||listar')
-    Permission.objects.get(name='muestra||ver')
-    Permission.objects.get(name='muestra||solicitar')
+    Permission.objects.get(name = 'Maquina||ver')
+    Permission.objects.get(name = 'Maquina||solicitar')
+    Permission.objects.get(name = 'usuario||agregar')
+    Permission.objects.get(name = 'muestra||listar')
+    Permission.objects.get(name = 'muestra||ver')
+    Permission.objects.get(name = 'muestra||solicitar')
 
     cientificos.permissions.add(
-        can_addProject,
-        can_editProject,
-        can_listProject,
-        can_viewProject,
-        can_addExperiment,
-        can_editExperiment,
-        can_listExperiment,
-        can_viewExperiment,
-        can_addUser,
-        can_editUser,
-        can_listUser,
-        can_viewUser,
-        can_addLab,
-        can_editLab,
-        can_listLab,
-        can_viewLab,
-        can_addMachine,
-        can_editMachine,
-        can_listMachine,
-        can_viewMachine,
-        can_addStorage,
-        can_editStorage,
-        can_listStorage,
-        can_viewStorage,
-        can_addProtocol,
-        can_editProtocol,
-        can_listProtocol,
-        can_viewProtocol,
-        can_addStep,
-        can_editStep,
-        can_listStep,
-        can_viewStep,
-        can_addSample,
-        can_editSample,
-        can_listSample,
-        can_viewSample,
-        can_listRequest,
-        can_viewRequest,
-        can_addTray,
-        can_editTray,
-        can_listTray,
-        can_viewTray,
+            can_addProject,
+            can_editProject,
+            can_listProject,
+            can_viewProject,
+            can_addExperiment,
+            can_editExperiment,
+            can_listExperiment,
+            can_viewExperiment,
+            can_addUser,
+            can_editUser,
+            can_listUser,
+            can_viewUser,
+            can_addLab,
+            can_editLab,
+            can_listLab,
+            can_viewLab,
+            can_addMachine,
+            can_editMachine,
+            can_listMachine,
+            can_viewMachine,
+            can_addStorage,
+            can_editStorage,
+            can_listStorage,
+            can_viewStorage,
+            can_addProtocol,
+            can_editProtocol,
+            can_listProtocol,
+            can_viewProtocol,
+            can_addStep,
+            can_editStep,
+            can_listStep,
+            can_viewStep,
+            can_addSample,
+            can_editSample,
+            can_listSample,
+            can_viewSample,
+            can_listRequest,
+            can_viewRequest,
+            can_addTray,
+            can_editTray,
+            can_listTray,
+            can_viewTray,
     )
     jefes.permissions.add(
-        can_listRequest,
-        can_viewRequest,
-        can_manageRequest,
-        can_listMachine,
-        can_viewMachine,
-        can_editMachine,
-        can_listStorage,
-        can_viewStorage,
-        can_editStorage,
-        can_addSample,
-        can_editSample,
-        can_listSample,
-        can_viewSample,
-        can_addUser,
+            can_listRequest,
+            can_viewRequest,
+            can_manageRequest,
+            can_listMachine,
+            can_viewMachine,
+            can_editMachine,
+            can_listStorage,
+            can_viewStorage,
+            can_editStorage,
+            can_addSample,
+            can_editSample,
+            can_listSample,
+            can_viewSample,
+            can_addUser,
     )
     asistentes.permissions.add(
-        can_listMachine,
-        can_viewMachine,
-        can_requestMachine,
-        can_listSample,
-        can_viewSample,
-        can_requestSample,
+            can_listMachine,
+            can_viewMachine,
+            can_requestMachine,
+            can_listSample,
+            can_viewSample,
+            can_requestSample,
     )
     if created1 or created2 or created3:
         return 0
@@ -371,7 +380,7 @@ def createGroups():
 def crearAsistente(user, num, tipDocumento, asistentes):
     print ('Crear Asistentes'),
     exist_asistente, new_asistente = User.objects.get_or_create(
-        username=user + str(num))
+            username = user + str(num))
     print ('...'),
     if new_asistente:
         exist_asistente.email = user + str(num) + '@uniandes.edu.co'
@@ -397,20 +406,19 @@ def crearAsistente(user, num, tipDocumento, asistentes):
 
 def createUsers():
     print ('Crear Usuarios'),
-    cientificos = Group.objects.get(name='Cientifico Experimentado')
-    jefes = Group.objects.get(name='Jefe de Laboratorio')
-    asistentes = Group.objects.get(name='Asistente de Laboratorio')
+    cientificos = Group.objects.get(name = 'Cientifico Experimentado')
+    jefes = Group.objects.get(name = 'Jefe de Laboratorio')
+    asistentes = Group.objects.get(name = 'Asistente de Laboratorio')
 
-    tipDocumento = TipoDocumento.objects.get(nombre_corto='CC')
+    tipDocumento = TipoDocumento.objects.get(nombre_corto = 'CC')
 
-    exist_cientifico, new_cientifico = User.objects.get_or_create(username='acastro')
+    exist_cientifico, new_cientifico = User.objects.get_or_create(username = 'acastro')
 
     if new_cientifico:
         exist_cientifico.email = 'acastro@uniandes.edu.co'
         exist_cientifico.set_password(CONTRASENA)
         exist_cientifico.groups.add(cientificos)
         exist_cientifico.save()
-
 
     Usuario.objects.get_or_create(
             nombre_usuario = 'acastro',
@@ -426,13 +434,12 @@ def createUsers():
             contrasena = CONTRASENA,
     )
     print ('.'),
-    exist_jefe, new_jefe = User.objects.get_or_create(username='bcamelas')
+    exist_jefe, new_jefe = User.objects.get_or_create(username = 'bcamelas')
     if new_jefe:
         exist_jefe.email = 'bcamelas@uniandes.edu.co'
         exist_jefe.set_password(CONTRASENA)
         exist_jefe.groups.add(jefes)
         exist_jefe.save()
-
 
     Usuario.objects.get_or_create(
             nombre_usuario = 'bcamelas',
@@ -497,4 +504,4 @@ class Command(BaseCommand):
                              ]
 
         for functions in lista_instalacion:
-            execute_function(self, function_name=functions[0], result_text=functions[1])
+            execute_function(self, function_name = functions[0], result_text = functions[1])
